backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
httpx==0.26.0
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.1
lxml==5.1.0
pikpakapi==0.1.11
pytz==2025.2
apscheduler==3.11.0
sqlalchemy==2.0.41
loguru==0.7.3